refactor providers
The current implementation based on an enum is not very flexible, It
basically avoids that one can write an implementation of another
provider in a separate plugin. And one has to change a lot of places to
create a new provider.
This change makes the provider an ExtensionPoint. This has some
advantages like provider specific requirements checking. E.g. the apiKey
is mandatory for OPOenAI and Gemini but not required for Ollama. And for
Ollama a url is mandatory.
The code is also a lot simpler and avoids special checks. Providers
implement checks if the config is valid avoiding unnecessary calls to
the provider when we know the request will fail anyway, e.g. due to
a missing apikey or a missing model.
This would also make it easily possible to change the plugin to have
multiple providers configured, either from different providers or
different modeld from the same provider.